[OPADF-1] Moved integration test
