+++ .travis.yml
@@ -41,10 +41,10 @@ jobs:
         - docker ps -a
       script:
         - ./gradlew clean build
-        - ./gradlew integrationTest
         - ./gradlew jacocoTestReport sonarqube -Dsonar.verbose=true
         - cd opa-datafilter-core
         - ./gradlew clean build
+        - ./gradlew integrationTest
         - ./gradlew jacocoTestReport sonarqube -Dsonar.verbose=true
       after_script:
         - docker ps -a
