+++ build.gradle
@@ -15,7 +15,6 @@ plugins {
 	id 'maven-publish'
 	id 'jacoco'
 	id "org.sonarqube" version "2.7"
-	id 'org.unbroken-dome.test-sets' version '2.2.1'
 	id 'io.codearte.nexus-staging' version '0.11.0'
 	id "io.franzbecker.gradle-lombok" version "1.14"
 }
@@ -75,31 +74,12 @@ dependencies {
 		exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
 	}
 	testImplementation 'de.flapdoodle.embed:de.flapdoodle.embed.mongo'
-	testImplementation 'org.mariadb.jdbc:mariadb-java-client:2.3.0'
-	testImplementation group: 'org.postgresql', name: 'postgresql', version: '42.2.10'
-	testImplementation group: 'org.testcontainers', name: 'mysql', version: '1.12.5'
-	testImplementation group: 'org.testcontainers', name: 'testcontainers', version: '1.12.5'
 	testImplementation group: 'com.h2database', name: 'h2', version: '1.4.197'
 	testImplementation group: 'org.hamcrest', name: 'hamcrest-all', version: '1.3'
 }
 
-testSets {
-	integrationTest
-}
-
 test {
 	useJUnitPlatform()
-}
-
-integrationTest {
-	useJUnitPlatform()
-	integrationTest.outputs.upToDateWhen { false }
-	failFast = true
-	dependsOn 'cleanIntegrationTest'
-
-	testLogging {
-		events 'passed', 'skipped', 'failed', 'standardOut', 'standardError'
-	}
 }
 
 sonarqube {
