+++ opa-datafilter-core/build.gradle
@@ -56,6 +56,10 @@ dependencies {
         exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
     }
 
+    testImplementation 'org.mariadb.jdbc:mariadb-java-client:2.3.0'
+    testImplementation group: 'org.postgresql', name: 'postgresql', version: '42.2.10'
+    testImplementation group: 'org.testcontainers', name: 'mysql', version: '1.12.5'
+    testImplementation group: 'org.testcontainers', name: 'testcontainers', version: '1.12.5'
     testImplementation group: 'com.h2database', name: 'h2', version: '1.4.197'
     testImplementation group: 'org.hamcrest', name: 'hamcrest-all', version: '1.3'
 }
