+++ opa-datafilter-core/src/main/java/opa/datafilter/core/ast/db/query/PredicateConverter.java
@@ -11,6 +11,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
+ * @author joffryferrater
+ *
  * Converts Open Policy Agent Abstract Syntax Tree predicate into SQL predicate.
  */
 public class PredicateConverter {
