+++ opa-datafilter-core/src/main/java/opa/datafilter/core/ast/db/query/util/SqlUtil.java
@@ -5,6 +5,9 @@ import opa.datafilter.core.ast.db.query.elements.SqlPredicate;
 
 import java.util.List;
 
+/**
+ * @author joffryferrater
+ */
 public class SqlUtil {
 
     private SqlUtil() {
