--- src/integrationTest/resources/Dockerfile_OPA
@@ -1,10 +1,0 @@
-FROM ubuntu:18.04
-
-RUN apt-get update && apt-get install -y curl
-RUN curl -L -o opa https://openpolicyagent.org/downloads/latest/opa_linux_amd64
-RUN chmod 755 ./opa
-
-COPY petclinic.rego .
-COPY run_opa.sh .
-
-ENTRYPOINT ["./run_opa.sh"]
