--- src/integrationTest/resources/pets-clinic-partial-request.json
@@ -1,12 +1,0 @@
-{
-  "query": "data.petclinic.authz.allow = true",
-  "input": {
-    "method": "GET",
-    "path": ["pets", "fluffy"],
-    "subject": {
-      "user": "alice",
-      "location": "SOMA"
-    }
-  },
-  "unknowns": ["data.pets"]
-}
