--- src/integrationTest/resources/run_opa.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-echo "Running open policy agent.."
-./opa run --server ./petclinic.rego
